Moved TtyBridge and TtyState to runtime
