+++ lib/wasix/src/lib.rs
@@ -85,16 +85,13 @@ use wasmer_wasix_types::wasi::{Errno, ExitCode};
 
 pub use crate::{
     fs::{default_fs_backing, Fd, WasiFs, WasiInodes, VIRTUAL_ROOT_FD},
-    os::{
-        task::{
-            control_plane::WasiControlPlane,
-            process::{WasiProcess, WasiProcessId},
-            thread::{WasiThread, WasiThreadError, WasiThreadHandle, WasiThreadId},
-        },
-        WasiTtyState,
+    os::task::{
+        control_plane::WasiControlPlane,
+        process::{WasiProcess, WasiProcessId},
+        thread::{WasiThread, WasiThreadError, WasiThreadHandle, WasiThreadId},
     },
     rewind::*,
-    runtime::{task_manager::VirtualTaskManager, PluggableRuntime, Runtime},
+    runtime::{task_manager::VirtualTaskManager, PluggableRuntime, Runtime, TtyState},
     state::{
         WasiEnv, WasiEnvBuilder, WasiEnvInit, WasiFunctionEnv, WasiInstanceHandles,
         WasiStateCreationError, ALL_RIGHTS,
